Check for existing .vim and .vimrc before symlinking

+++ setup.sh
@@ -9,8 +9,21 @@ git submodule update
 
 # Symlink vimrc & vim into your home directory (i.e. ~/.vimrc and ~/.vim)
 cd
-ln -s $dir/vimrc .vimrc
-ln -s $dir/vim .vim
+
+if [ ! -L .vimrc ]; then
+  if [ -e .vimrc ]; then
+    mv .vimrc .vimrc.bak
+  fi
+  ln -s $dir/vimrc .vimrc
+fi
+
+if [ ! -L .vim ]; then
+  if [ -e .vim ]; then
+    mv .vim .vim.bak
+  fi
+  ln -s $dir/vim .vim
+fi
+
 
 # Install plugins
 echo
